fix: use i32::MAX to avoid block_number overflow in SQL query
The block_number parameter was passed as u32::MAX which overflows to -1
when cast to i32 in the SQL query, causing no rows to match.

+++ node/src/aiken_authority_selection.rs
@@ -77,7 +77,8 @@ where
 		_permissioned_candidates_policy: PolicyId,
 	) -> Result<AriadneParameters, Box<dyn std::error::Error + Send + Sync>> {
 		// Use Aiken parser directly for permissioned candidates
-		let candidates = self.get_aiken_permissioned_candidates(u32::MAX).await?;
+		// Note: Use i32::MAX to avoid overflow when cast to i32 in SQL query
+		let candidates = self.get_aiken_permissioned_candidates(i32::MAX as u32).await?;
 
 		log::info!(
 			"Aiken parser found {} permissioned candidates for epoch {}",
